Outsource `set_dot_basics()` and `apply_dot_tweaks()`
